Assets/StreamingAssets/pyproject.toml
[tool.poetry]
name = "streamingassets"
version = "0.1.0"
description = ""
authors = ["jmigsan <43961001+jmigsan@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
python-multipart = "^0.0.20"
numba = "^0.61.0"
openai-whisper = "^20240930"
torch = {version = "^2.6.0+cu118", source = "pytorch-gpu"}
torchvision = {version = "^0.21.0+cu118", source = "pytorch-gpu"}
torchaudio = {version = "^2.6.0+cu118", source = "pytorch-gpu"}
edge-tts = "^7.0.0"
numpy = "^2.1"
pydantic = "^2.10.6"

[[tool.poetry.source]]
name = "pytorch-gpu"
url = "https://download.pytorch.org/whl/cu118"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
